Added routing, and a first attempt at the history page
