+++ src/frontend/notifai-uploader/package.json
@@ -25,6 +25,7 @@
     "react-dom": "^19.1.0",
     "react-markdown": "^10.1.0",
     "react-oidc-context": "^3.3.0",
+    "react-router-dom": "^7.8.0",
     "react-scripts": "5.0.1",
     "spinners-react": "^1.0.11",
     "web-vitals": "^2.1.4"
