+++ src/frontend/notifai-uploader/public/env-config.js
@@ -1,0 +1,1 @@
+window.env = {};
