+++ src/frontend/notifai-uploader/src/components/Header.js
@@ -1,6 +1,7 @@
 import './Header.css';
 import './Shared.css';
 import { useAuth } from './AuthContext';
+import { Link } from "react-router-dom";
 
 export default function Header() {
   const { logout } = useAuth();
@@ -8,6 +9,16 @@ export default function Header() {
     <header>
       <img src={window.env?.PUBLIC_URL || process.env.PUBLIC_URL + '/nhs-england-white.svg'} alt="NHS logo" className="nhsuk-header-logo" />
       <h1>Notify AI</h1>
+      <nav>
+        <ul className="nhsuk-header__navigation" style={{ display: 'flex', justifyContent: 'center', marginLeft: '50px' }}>
+          <li className="nhsuk-header__navigation-item" style={{ paddingLeft: '15px' }}>
+            <Link to="/" className="nhsuk-header__navigation-link">File Upload</Link>
+          </li>
+          <li className="nhsuk-header__navigation-item" style={{ paddingLeft: '15px' }}>
+            <Link to="/history" className="nhsuk-header__navigation-link">History</Link>
+          </li>
+        </ul>
+      </nav>
       <button
         onClick={logout}
         className="nhsuk-button nhsuk-button--secondary"
