+++ src/frontend/notifai-uploader/src/components/History.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+function History() {
+  return (
+    <div>
+      <main className="container">
+        <div className="two-column-content">
+          <h1>History of File Uploads</h1>
+          <p>This page will display the history of file uploads.</p>
+        </div>
+      </main>
+    </div>
+  );
+}
+
+export default History;
