+++ src/frontend/notifai-uploader/src/index.js
@@ -11,7 +11,7 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <AuthProvider>
-      <App/>
+        <App/>
     </AuthProvider>
   </React.StrictMode>
 );
